--- app/lib/biblioteca_liliac.py
@@ -1,11 +1,0 @@
-def descriere_liliac():
-    return "Liliacul (Syringa vulgaris) face parte din familia oleaceelor și este o specie din genul Syringa care înflorește primăvara. Este un arbust a cărui înălțime poate ajunge până la șapte metri, cu ramuri drepte și lujeri puțin muchiați."
-
-def frunze_liliac():
-    return "Frunzele sunt ovate sau lat-ovate, la bază cordate, până la 12 centimetri lungime, acuminate, pețiol circa 2,5 cm lungime, glabre, verzi-întunecat."
-
-def floare_liliac():
-    return "Florile sunt simple sau duble, plăcut mirositoare, grupate în penicule de până la 20 cm lungime, multiflorale, în culori diferite, de la liliachiu la alb. Floarea are caliciul mic, campanulat, cu patru dinți, corolă cu tub de un centimetru lungime, cu patru lobi patenți și două stamine."
-
-def culoare_liliac():
-    return "Lila este o culoare asemănătoare cu violetul, dar o nuanță mai deschisă, spre roz. Numele provine de la florile de liliac, care au această culoare."
