--- app/templates/frunze.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    
-    <h1>Frunze</h1>
-    <h2>{{frunze}}</h2>
-
-</body>
-</html>
