--- app/templates/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Liliac</title>
-    <link rel="stylesheet" href="../style/style.css">
-</head>
-<body>
-
-    <h1>Descriere liliac</h1>
-    <h2>{{descriere}}</h2>
-</body>
-</html>
